requirements.txt
Flask
pytesseract
pillow
opencv-python
numpy
pandas
